requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==2.2.5
Django==4.1.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-grappelli==3.0.3
django-jazzmin==2.5.0
django-paypal==0.2.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gyp==0.1
html5lib==1.1
httpie==3.2.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.2
mysql-connector-python==8.0.30
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==12.6.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.1
South==1.0.2
sqlparse==0.4.2
ssh-import-id==5.11
stripe==4.2.0
style==1.1.0
sympy==1.9
systemd-python==234
traitlets==5.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
update==0.0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
